# [BAD] refactor(gnovm): handle assignment operations in preprocess