--- a/gnovm/pkg/gnolang/nodes.go
+++ b/gnovm/pkg/gnolang/nodes.go
@@ -116,6 +116,8 @@ const (
 	maxAssignmentOperation = BAND_NOT_ASSIGN
 )
 
+// isAssignmentOperation determines if the Word is an assignment operation;
+// ie, a combination of an assignment and a binary operator like ADD_ASSIGN.
 func (w Word) isAssignmentOperation() bool {
 	return w >= minAssignmentOperation && w <= maxAssignmentOperation
 }
@@ -126,7 +128,6 @@ func (w Word) isAssignmentOperation() bool {
 //
 // If w is not an assignment operation, ILLEGAL is returned.
 func (w Word) convertAssignmentOperation() Word {
-	// XXX: is this inlined? if not, would lookup table make it inlined / better?
 	switch w {
 	case ADD_ASSIGN:
 		return ADD