… gnovm/pkg/gnolang/op_assign.go